requirements.txt
# ApplyPotato Dependencies

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Google APIs
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# AI providers
openai>=1.6.0
google-genai>=1.0.0

# Scheduling
apscheduler>=3.10.4

# Configuration
python-dotenv>=1.0.0

# HTTP requests (for GitHub markdown fetching)
httpx>=0.25.0

# Logging
colorlog>=6.8.0

# Fuzzy string matching
rapidfuzz>=3.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
